Attack rework + Broken Code
Reworked attack to electric themed
changed it to a spam based attack, not relying on a cooldown to use. also fixed the sprite issue by doing away with the timer as well. Somehow though i managed to break the falling blocks colision, now i can jump on the block even if it has not respawned